+++ src/containers/Profile/ProfileStyles/Forms.css
@@ -13,8 +13,8 @@
   padding: 3px 5px;
 }
 .profile-tabs {
-  display: flex;
-  flex-wrap: wrap;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
   justify-content: space-between;
   margin-bottom: 30px;
 }
